chore: bump Nx to 19.3.2
