+++ .eslintrc.json
@@ -20,7 +20,12 @@
         "prefer": "type-imports"
       }
     ],
-    "import/no-extraneous-dependencies": ["error", { "devDependencies": false }],
+    "import/no-extraneous-dependencies": [
+      "error",
+      {
+        "devDependencies": false
+      }
+    ],
     "jest/no-focused-tests": "error"
   },
   "overrides": [
@@ -49,13 +54,18 @@
       "rules": {
         "@typescript-eslint/no-inferrable-types": "off",
         "@typescript-eslint/no-non-null-assertion": "off",
-        "@typescript-eslint/no-unused-vars": "error"
+        "@typescript-eslint/no-unused-vars": "error",
+        "@typescript-eslint/no-extra-semi": "error",
+        "no-extra-semi": "off"
       }
     },
     {
       "files": ["*.js", "*.jsx"],
       "extends": ["plugin:@nx/javascript"],
-      "rules": {}
+      "rules": {
+        "@typescript-eslint/no-extra-semi": "error",
+        "no-extra-semi": "off"
+      }
     },
     {
       "files": [
@@ -64,7 +74,8 @@
         "**/*.stories.tsx",
         "**/*.test.ts",
         "**/*.test.tsx",
-        "**/jest.setup.js"
+        "**/jest.setup.js",
+        "**/jest.setup.ts"
       ],
       "rules": {
         "import/no-extraneous-dependencies": "off"
