+++ .gitignore
@@ -45,3 +45,4 @@ testem.log
 Thumbs.db
 
 .nx/cache
+.nx/workspace-data
