+++ apps/benchmark/project.json
@@ -8,7 +8,7 @@
       "executor": "nx:run-commands",
       "options": {
         "cwd": "apps/benchmark",
-        "commands": [{ "command": "node ./build.js" }]
+        "commands": [{ "command": "node ./build.mjs" }]
       },
       "outputs": ["{workspaceRoot}/dist/apps/benchmark"]
     },
@@ -16,7 +16,7 @@
       "executor": "nx:run-commands",
       "options": {
         "cwd": "apps/benchmark",
-        "commands": [{ "command": "node ./build.js --watch" }],
+        "commands": [{ "command": "node ./build.mjs --watch" }],
         "outputPath": ["dist/apps/benchmark"]
       }
     },
