+++ change/@griffel-jest-serializer-2e0f1b09-c42e-43b9-9c5c-87b88239c2e5.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: bump Nx to 19.3.2",
+  "packageName": "@griffel/jest-serializer",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "none"
+}
