+++ package.json
@@ -36,25 +36,25 @@
     "@docusaurus/preset-classic": "^3.4.0",
     "@docusaurus/theme-mermaid": "^3.4.0",
     "@emotion/css": "^11.9.0",
-    "@nx/eslint": "18.3.5",
-    "@nx/eslint-plugin": "18.3.5",
-    "@nx/jest": "18.3.5",
-    "@nx/js": "18.3.5",
-    "@nx/node": "18.3.5",
-    "@nx/react": "18.3.5",
-    "@nx/rollup": "18.3.5",
-    "@nx/storybook": "18.3.5",
-    "@nx/web": "18.3.5",
-    "@nx/webpack": "18.3.5",
-    "@nx/workspace": "18.3.5",
+    "@nx/eslint": "19.3.2",
+    "@nx/eslint-plugin": "19.3.2",
+    "@nx/jest": "19.3.2",
+    "@nx/js": "19.3.2",
+    "@nx/node": "19.3.2",
+    "@nx/react": "19.3.2",
+    "@nx/rollup": "19.3.2",
+    "@nx/storybook": "19.3.2",
+    "@nx/web": "19.3.2",
+    "@nx/webpack": "19.3.2",
+    "@nx/workspace": "19.3.2",
     "@rspack/cli": "0.7.5",
     "@rspack/core": "0.7.5",
     "@storybook/addon-essentials": "7.6.8",
     "@storybook/core-server": "7.6.8",
     "@storybook/react-webpack5": "7.6.8",
-    "@swc/core": "1.3.102",
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "13.4.0",
+    "@swc/core": "1.5.29",
+    "@testing-library/jest-dom": "6.4.6",
+    "@testing-library/react": "15.0.6",
     "@types/babel__helper-module-imports": "7.18.0",
     "@types/babel__helper-plugin-utils": "7.10.0",
     "@types/chrome": "0.0.180",
@@ -66,8 +66,8 @@
     "@types/mini-css-extract-plugin": "2.5.1",
     "@types/node": "18.19.9",
     "@types/prettier": "2.7.3",
-    "@types/react": "18.2.24",
-    "@types/react-dom": "18.2.9",
+    "@types/react": "18.3.1",
+    "@types/react-dom": "18.3.0",
     "@types/styled-components": "5.1.26",
     "@types/stylis": "4.2.0",
     "@types/tmp": "0.2.3",
@@ -77,7 +77,6 @@
     "@typescript-eslint/parser": "7.14.1",
     "@uifabric/merge-styles": "7.19.1",
     "babel-jest": "29.4.3",
-    "babel-loader": "8.1.0",
     "babel-plugin-annotate-pure-calls": "0.4.0",
     "babel-plugin-tester": "11.0.0",
     "beachball": "2.31.5",
@@ -86,7 +85,7 @@
     "css-loader": "^6.4.0",
     "d3-scale-chromatic": "^3.0.0",
     "doctoc": "2.2.1",
-    "esbuild": "0.19.11",
+    "esbuild": "0.21.5",
     "esbuild-plugin-alias": "0.2.1",
     "esbuild-plugin-import-glob": "0.1.1",
     "eslint": "8.57.0",
@@ -117,13 +116,13 @@
     "monosize-storage-upstash": "0.0.8",
     "nano-staged": "0.5.0",
     "next": "14.1.1",
-    "nx": "18.3.5",
-    "postcss": "^8.4.31",
+    "nx": "19.3.2",
+    "postcss": "8.4.38",
     "prettier": "2.8.2",
     "prism-react-renderer": "2.3.1",
     "raw-loader": "4.0.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-fela": "12.2.0",
     "react-refresh": "^0.10.0",
     "react-shadow": "^20.2.0",
@@ -137,7 +136,7 @@
     "ts-node": "10.9.1",
     "tsconfig-paths": "4.1.1",
     "typescript": "5.4.5",
-    "vite": "^4.5.3",
+    "vite": "^5.3.2",
     "webpack": "5.89.0",
     "webpack-merge": "5.10.0",
     "yargs": "^17.5.1"
